Refactoring and cleanup. Added userauth and home app
