+++ hirearefugee/hirearefugee/settings.py
@@ -37,8 +37,10 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'useroffer',
+    'home',
     'support',
+    'userauth',
+    'useroffer',
 ]
 
 MIDDLEWARE = [
