+++ hirearefugee/hirearefugee/urls.py
@@ -15,16 +15,18 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from useroffer import views as useroffer_views
+from home import views as home_views
 from support import views as support_views
+from userauth import views as userauth_views
+from useroffer import views as useroffer_views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
 
     # Auth
-    path('signup/', useroffer_views.signupuser, name='signupuser'),
-    path('login/', useroffer_views.loginuser, name='loginuser'),
-    path('logout/', useroffer_views.logoutuser, name='logoutuser'),
+    path('userauth/signup/', userauth_views.signupuser, name='signupuser'),
+    path('userauth/login/', userauth_views.loginuser, name='loginuser'),
+    path('userauth/logout/', userauth_views.logoutuser, name='logoutuser'),
 
     # Support
     path('support/becomeamember', support_views.becomeamember, name='becomeamember'),
@@ -34,12 +36,13 @@ urlpatterns = [
     path('support/spreadtheword', support_views.spreadtheword, name='spreadtheword'),
     path('support/thanks', support_views.thanks, name='thanks'),
 
-    # Navbar
-    path('about/', useroffer_views.about, name='about'),
-    path('testhtml/', useroffer_views.testhtml, name='testhtml'),
+    # Home
+    path('', home_views.home, name='home'),
+    path('home/', home_views.home, name='home'),
+    path('home/about/', home_views.about, name='about'),
+    path('home/testhtml/', home_views.testhtml, name='testhtml'),
 
     # UserOffers
-    path('', useroffer_views.home, name='home'),
     path('create/', useroffer_views.createuseroffer, name='createuseroffer'),
     path('current/', useroffer_views.currentuseroffers, name='currentuseroffers'),
     path('completed/', useroffer_views.completeduseroffers, name='completeduseroffers'),
