+++ hirearefugee/home/templates/home/about.html
@@ -1,4 +1,4 @@
-{% extends "useroffer/base.html" %}
+{% extends "home/base.html" %}
 
 {% block content %}
 <br>
