+++ hirearefugee/home/templates/home/base.html
@@ -14,7 +14,7 @@
     <!-- Bootstrap CSS -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
 
-    <link rel="icon" type="image/png"  href="{% static 'useroffer/Shake_Hands_Icon_0ea8b3.png' %}">
+    <link rel="icon" type="image/png"  href="{% static 'home/icons/Shake_Hands_Icon_0ea8b3.png' %}">
     <title>Hire A Refugee</title>
 </head>
 <body>
@@ -23,7 +23,7 @@
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
 <div class="container">
     <a class="navbar-brand" href="{% url 'home' %}">
-        <img src="{% static 'useroffer/Shake_Hands_Icon_0ea8b3.png' %}" width="35" height="35" class="d-inline-block align-top">
+        <img src="{% static 'home/icons/Shake_Hands_Icon_0ea8b3.png' %}" width="35" height="35" class="d-inline-block align-top">
     </a>
     <a class="navbar-brand" href="{% url 'home' %}">
         <span>Hire A Refugee</span> <span style="color:rgb(14, 168, 179)">
@@ -62,7 +62,7 @@
                 <!-- Create -->
                 <li class="nav-item dropdown">
                     <a class="nav-link dropdown-toggle" href="#" id="navbarDropdownMenuLink" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                        <img src="{% static 'useroffer/Heart Dark_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
+                        <img src="{% static 'home/icons/Heart Dark_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
                     </a>
                     <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
                         <a class="dropdown-item" href="#">Help</a>
@@ -74,35 +74,35 @@
                 <!-- Map -->
                 <li class="nav-item">
                     <a href="#" onclick="$('#').click()" class="nav-link">
-                        <img src="{% static 'useroffer/Map location_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
+                        <img src="{% static 'home/icons/Map location_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
                     </a>
                 </li>
 
                 <!-- Home -->
                 <li class="nav-item">
                     <a href="#" onclick="$('#').click()" class="nav-link">
-                        <img src="{% static 'useroffer/Nice_House_Icon_0000ff.png' %}" width="20" height="20" class="d-inline-block align-top">
+                        <img src="{% static 'home/icons/Nice_House_Icon_0000ff.png' %}" width="20" height="20" class="d-inline-block align-top">
                     </a>
                 </li>
                 
                 <!-- Messaging -->
                 <li class="nav-item">
                     <a href="#" onclick="$('#').click()" class="nav-link">
-                        <img src="{% static 'useroffer/Mail Dark_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
+                        <img src="{% static 'home/icons/Mail Dark_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
                     </a>
                 </li>
 
                 <!-- Settings -->
                 <li class="nav-item">
                     <a href="#" onclick="$('#').click()" class="nav-link">
-                        <img src="{% static 'useroffer/Gear_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
+                        <img src="{% static 'home/icons/Gear_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
                     </a>
                 </li>
 
                 <!-- Logout -->
                 <li class="nav-item">
                     <a href="#" onclick="$('#signOutBtn').click()" class="nav-link">
-                        <img src="{% static 'useroffer/Door_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
+                        <img src="{% static 'home/icons/Door_000000.png' %}" width="20" height="20" class="d-inline-block align-top">
                     </a>
                     <form style='display: none;' method="POST" action="{% url 'logoutuser' %}">
                         {% csrf_token %}
@@ -118,12 +118,12 @@
                 </li>
                 <li class="nav-item mx-1 my-1">
                     <a href="https://www.instagram.com/hirearefugee/" target="_blank">
-                        <img src="{% static 'useroffer/instagram_bw.svg' %}" width="30" height="30" class="d-inline-block align-top">
+                        <img src="{% static 'home/icons/instagram_bw.svg' %}" width="30" height="30" class="d-inline-block align-top">
                     </a>
                 </li>
                 <li class="nav-item mx-1 my-1">
                     <a href="https://github.com/maximilianharr/hirearefugee" target="_blank">
-                        <img src="{% static 'useroffer/github_bw.svg' %}" width="30" height="30" class="d-inline-block align-top">
+                        <img src="{% static 'home/icons/github_bw.svg' %}" width="30" height="30" class="d-inline-block align-top">
                     </a>
                 </li>
             {% endif %}
