+++ hirearefugee/home/templates/home/home.html
@@ -1,10 +1,10 @@
-{% extends "useroffer/base.html" %}
+{% extends "home/base.html" %}
 
 {% block content %}
 <br>
-    <br>
-    <br>
-    <br>
+<br>
+<br>
+<br>
 <div class="row justify-content-center mt-5">
     <div class="col-md-7 text-center">
         <h1 style="color:rgb(14, 168, 179)">Hire A Refugee!</h1>
