+++ hirearefugee/support/templates/support/spreadtheword.html
@@ -1,7 +1,61 @@
-{% extends "useroffer/base.html" %}
+{% extends "home/base.html" %}
 
 {% block content %}
 
-Spread The Word
+{% load static %}
 
+<div class="row justify-content-center mt-5">
+    <div class="col-md-7 text-center">
+        <h1 style="color:rgb(14, 168, 179)">Feel Free To Share Our QR-Codes!</h1>
+    </div>
+</div>
+
+<div class="album py-5">
+    <!-- First item -->
+    <div class="container">
+        <div class="row">
+            <div class="col-md-4">
+                <div class="card mb-4 shadow-sm">
+                <img src="{% static 'home/qr/http_www_hirearefugee_org_.png' %}" width="100%">
+                    <div class="card-body">
+                    <p class="card-text">
+                        <img src="{% static 'home/icons/Shake_Hands_Icon_0ea8b3.png' %}" width="50"> Website <br>
+                        https://hirearefugee.org/
+                    </p>
+                    <div class="d-flex justify-content-between align-items-center">
+                    </div>
+                </div>
+                </div>
+            </div>
+            <!-- First item -->
+            <div class="col-md-4">
+                <div class="card mb-4 shadow-sm">
+                    <img src="{% static 'home/qr/https_www_github_com_maximilianharr_hirearefugee_.png' %}" width="100%">
+                    <div class="card-body">
+                    <p class="card-text"> 
+                        <img src="{% static 'home/icons/github_bw.svg' %}" width="50"> Github <br>
+                        https://github.com/maximilianharr/hirearefugee
+                    </p>
+                    <div class="d-flex justify-content-between align-items-center">
+                    </div>
+                    </div>
+                </div>
+            </div>
+            <!-- First item -->
+            <div class="col-md-4">
+                <div class="card mb-4 shadow-sm">
+                    <img src="{% static 'home/qr/https_www_instagram_com_hirearefugee_.png' %}" width="100%">
+                    <div class="card-body">
+                    <p class="card-text">
+                        <img src="{% static 'home/icons/instagram_bw.svg' %}" width="50"> Instagram <br>
+                        https://www.instagram.com/hirearefugee/
+                    </p>
+                    <div class="d-flex justify-content-between align-items-center">
+                    </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+  </div>
 {% endblock %}
