+++ hirearefugee/useroffer/templates/useroffer/createuseroffer.html
@@ -1,4 +1,4 @@
-{% extends "useroffer/base.html" %}
+{% extends "home/base.html" %}
 
 {% block content %}
     <div class="row justify-content-center mt-5">
