+++ hirearefugee/useroffer/templates/useroffer/viewuseroffer.html
@@ -4,7 +4,7 @@
 
 
 
-{% extends "useroffer/base.html" %}
+{% extends "home/base.html" %}
 
 {% block content %}
     <div class="row justify-content-center mt-5">
